Define Rooms, Actions, and Triggers
Rooms can have multiple actions but only a single trigger. Actions are the allowed moves a hero can choose from after entering a room. The only action we currently have in the game is moving forward.

Once a hero selects an action, that action is used to configure the trigger. Given the hero that entered the room and the action that the hero chose, the trigger kicks off the events that will happen next.

Note that the two triggers in the game, Exit and Enemy, both implement the Trigger behaviour. Elixir behaviours allow you to define an interface that will be shared across modules.